requirements.txt
# Core ML/DL frameworks
torch>=1.9.0
torchvision>=0.10.0
pytorch-lightning>=1.5.0

# Data processing
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
scipy>=1.7.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Statistical analysis
statsmodels>=0.12.0

# Configuration and utilities
PyYAML>=5.4.0
tqdm>=4.62.0

# Experiment tracking
wandb>=0.12.0

# Medical imaging (optional, for NIfTI support)
nibabel>=3.2.0
SimpleITK>=2.1.0

# Additional utilities
pillow>=8.3.0
opencv-python>=4.5.0
einops>=0.6.0

# GUI for attention map viewer
PySide6>=6.0.0
